feat: AppHeaderBar
