+++ lib/dashboard.dart
@@ -1,12 +1,10 @@
 import 'dart:async';
-import 'dart:collection';
 import 'dart:io';
 
 import 'package:flutter/material.dart';
-import 'package:lux/tr.dart';
 import 'package:lux/utils.dart';
 import 'package:lux/widget/app_bottom_bar.dart';
-import 'package:url_launcher/url_launcher.dart';
+import 'package:lux/widget/app_header_bar.dart';
 import 'package:window_manager/window_manager.dart';
 
 import 'core_manager.dart';
@@ -25,15 +23,13 @@ class Dashboard extends StatefulWidget {
 }
 
 class _DashboardState extends State<Dashboard> with WindowListener {
-  bool isStarted = false;
-  String curProxyInfo = "";
   ProxyList proxyList = ProxyList(<ProxyItem>[], "");
-  RuleList ruleList = RuleList(<String>[], "");
-  bool isLoadingSwitch = false;
+
+  String curProxyInfo = "";
+
   bool isLoadingProxyList = false;
-  bool isLoadingRuleList = false;
+
   bool isLoadingProxyRadio = false;
-  bool isLoadingRuleDropdown = false;
 
   _DashboardState();
 
@@ -70,75 +66,11 @@ class _DashboardState extends State<Dashboard> with WindowListener {
     refreshData();
   }
 
-  Future<void> refreshIsStarted() async {
-    final value = await widget.coreManager.getIsStarted();
-    setState(() {
-      isStarted = value;
-    });
-  }
-
-  Future<void> refreshCurProxyInfo() async {
-    final value = await widget.coreManager.getCurProxyInfo();
-    setState(() {
-      curProxyInfo = value;
-    });
-  }
-
   Future<void> refreshProxyList() async {
     final value = await widget.coreManager.getProxyList();
     setState(() {
       proxyList = value;
     });
-  }
-
-  Future<void> refreshRuleList() async {
-    final value = await widget.coreManager.getRuleList();
-    setState(() {
-      ruleList = value;
-    });
-  }
-
-  Future<void> refreshData() async {
-    if (!isLoadingSwitch) {
-      refreshIsStarted();
-    }
-
-    refreshCurProxyInfo();
-
-    if (!isLoadingProxyRadio) {
-      refreshProxyList();
-    }
-
-    if (!isLoadingRuleDropdown) {
-      refreshRuleList();
-    }
-  }
-
-  void onSwitchChanged(bool value) async {
-    try {
-      setState(() {
-        isLoadingSwitch = true;
-      });
-      if (value) {
-        await widget.coreManager.start();
-        setState(() {
-          isStarted = true;
-        });
-      } else {
-        await widget.coreManager.stop();
-        setState(() {
-          isStarted = false;
-        });
-      }
-    } finally {
-      setState(() {
-        isLoadingSwitch = false;
-      });
-    }
-  }
-
-  void openWebDashboard() {
-    launchUrl(Uri.parse(widget.urlStr));
   }
 
   Future<void> handleSelectProxy(String? id) async {
@@ -164,90 +96,28 @@ class _DashboardState extends State<Dashboard> with WindowListener {
     }
   }
 
-  Future<void> handleSelectRule(String? id) async {
-    if (id == null) {
-      return;
-    }
-    try {
-      setState(() {
-        isLoadingRuleDropdown = true;
-      });
-      await widget.coreManager.selectRule(id);
-      setState(() {
-        ruleList.selectedId = id;
-      });
-    } finally {
-      setState(() {
-        isLoadingRuleDropdown = false;
-      });
+  Future<void> refreshData() async {
+    if (!isLoadingProxyRadio) {
+      refreshProxyList();
     }
   }
 
-  String getRuleLabel(String name) {
-    switch (name) {
-      case "proxy_all":
-        return tr().proxyAllRuleLabel;
-      case "proxy_gfw":
-        return tr().proxyGFWRuleLabel;
-      case "bypass_cn":
-        return tr().bypassCNRuleLabel;
-      case "bypass_all":
-        return tr().bypassAllRuleLabel;
-      default:
-        return name;
-    }
+  void onCurProxyInfoChange(String info) {
+    setState(() {
+      curProxyInfo = info;
+    });
   }
 
   @override
   Widget build(BuildContext context) {
-    final List<DropdownMenuEntry<String>> menuEntries =
-        UnmodifiableListView<DropdownMenuEntry<String>>(
-      ruleList.rules.map<DropdownMenuEntry<String>>((String name) {
-        String label = getRuleLabel(name);
-        return DropdownMenuEntry(value: name, label: label);
-      }),
-    );
-
     return Scaffold(
-      appBar: AppBar(
-          leading: IconButton(
-              tooltip: tr().goWebDashboardTip,
-              onPressed: openWebDashboard,
-              padding: EdgeInsetsGeometry.all(0),
-              icon: const Icon(
-                Icons.settings,
-              )),
-          title: Row(
-            children: [
-              SizedBox(
-                height: 32,
-                child: FittedBox(
-                  child: DropdownMenu<String>(
-                    width: 180,
-                    initialSelection: ruleList.selectedId,
-                    onSelected: isLoadingRuleDropdown ? null : handleSelectRule,
-                    dropdownMenuEntries: menuEntries,
-                    textStyle: TextStyle(fontSize: 20),
-                  ),
-                ),
-              ),
-              SizedBox(width: 4),
-              Spacer(),
-              Text(
-                curProxyInfo,
-                style: TextStyle(fontSize: 14),
-              ),
-              SizedBox(width: 8),
-              SizedBox(
-                width: 48,
-                child: FittedBox(
-                  child: Switch(
-                    value: isStarted,
-                    onChanged: isLoadingSwitch ? null : onSwitchChanged,
-                  ),
-                ),
-              ),
-            ],
+      appBar: PreferredSize(
+          preferredSize: Size.fromHeight(100.0),
+          child: AppHeaderBar(
+            coreManager: widget.coreManager,
+            urlStr: widget.urlStr,
+            curProxyInfo: curProxyInfo,
+            onCurProxyInfoChange: onCurProxyInfoChange,
           )),
       body: ListView(
         children: [
